Move getMappingObjectKind() back to ComposedMetricMapping
The manager is not always available and the method is easier to find on the ComposedMetricMapping than as a stand alone utility function
