+++ ts/libs/core/src/lib/composed-metrics/public/common/index.ts
@@ -1,5 +1,4 @@
 export * from './composed-metric-source-factory';
 export * from './composed-metric-source.base';
 export * from './composed-metric-source';
-export * from './composed-metric-type';
 export * from './errors';
