+++ ts/libs/core/src/lib/composed-metrics/public/control/composed-metric-mapping-manager.ts
@@ -1,5 +1,4 @@
-import { ComposedMetricMapping, ComposedMetricParams, ObjectKind  } from '../../../model';
-import { ComposedMetricType } from '../common';
+import { ComposedMetricMapping, ComposedMetricParams, ComposedMetricType  } from '../../../model';
 
 /**
  * Used to check for and configure `ComposedMetricMappings` in the orchestrator.
@@ -18,14 +17,5 @@ export interface ComposedMetricMappingManager {
         metricType: ComposedMetricType<V, P>,
         params: P,
     ): Promise<ComposedMetricMapping>;
-
-    /**
-     * Gets the {@link ObjectKind} for the `ComposedMetricMapping` for the specified `metricType`.
-     *
-     * @param metricType The {@link ComposedMetricType}, for which to get the `ObjectKind`.
-     * @param override (optional) Provides the possibility to override one or more `ObjectKind` properties with custom values.
-     * @returns A new {@link ObjectKind} instance.
-     */
-    getMappingObjectKind(metricType: ComposedMetricType<any>, override?: Partial<ObjectKind>): ObjectKind;
 
 }
