+++ ts/libs/core/src/lib/model/composed-metric-type.ts
@@ -1,5 +1,5 @@
-import { ComposedMetricParams } from '../../../model';
-import { TypeFn } from '../../../util';
+import { TypeFn } from '../util';
+import { ComposedMetricParams } from './composed-metric-params.prm';
 
 /**
  * Describes a type of composed metric.
