+++ ts/libs/core/src/lib/runtime/public/metrics-source/metrics-source.ts
@@ -1,5 +1,5 @@
-import { ComposedMetricSource, ComposedMetricType } from '../../../composed-metrics';
-import { ComposedMetricParams } from '../../../model';
+import { ComposedMetricSource } from '../../../composed-metrics';
+import { ComposedMetricParams, ComposedMetricType } from '../../../model';
 import { TimeSeriesSource } from '../../../raw-metrics-query/public/time-series';
 
 /**
